@@ -11,29 +11,27 @@
   },
   "private": true,
   "dependencies": {
-    "@angular-devkit/build-angular": "^14.2.7",
-    "@angular/animations": "^14.2.8",
+    "@angular-devkit/build-angular": "^15.2.7",
+    "@angular/animations": "^15.2.8",
     "@angular/cdk": "^14.2.6",
-    "@angular/cli": "^14.2.10",
-    "@angular/common": "^14.2.8",
-    "@angular/compiler": "^14.2.8",
-    "@angular/compiler-cli": "^14.2.8",
-    "@angular/core": "^14.2.8",
-    "@angular/flex-layout": "^14.0.0-beta.41",
-    "@angular/forms": "^14.2.8",
+    "@angular/cli": "^15.2.7",
+    "@angular/common": "^15.2.8",
+    "@angular/compiler": "^15.2.8",
+    "@angular/compiler-cli": "^15.2.8",
+    "@angular/core": "^15.2.8",
+    "@angular/forms": "^15.2.8",
     "@angular/material": "^14.2.6",
     "@angular/material-moment-adapter": "^14.2.6",
-    "@angular/platform-browser": "^14.2.8",
-    "@angular/platform-browser-dynamic": "^14.2.8",
-    "@angular/router": "^14.2.8",
+    "@angular/platform-browser": "^15.2.8",
+    "@angular/platform-browser-dynamic": "^15.2.8",
+    "@angular/router": "^15.2.8",
     "@babel/core": "^7.14.2",
     "@formatjs/intl-displaynames": "^6.2.6",
     "@fortawesome/angular-fontawesome": "^0.10.2",
     "@fortawesome/fontawesome-svg-core": "^1.2.35",
     "@fortawesome/free-solid-svg-icons": "^5.15.3",
-    "@ngrx/effects": "^14.3.2",
-    "@ngrx/store": "^14.3.2",
-    "@ngrx/store-devtools": "^14.3.2",
+    "@ngrx/effects": "^15.4.0",
+    "@ngrx/store": "^15.4.0",
     "@ngx-translate/core": "^14.0.0",
     "@ngx-translate/http-loader": "^7.0.0",
     "@turf/boolean-point-in-polygon": "^6.5.0",
@@ -46,7 +44,6 @@
     "@yellowspot/ng-truncate": "^2.0.0",
     "ajv": "^6.12.6",
     "angular-resizable-element": "^5.0.0",
-    "angular-resize-event": "^3.2.0",
     "braces": "^3.0.2",
     "cookieconsent": "^3.1.1",
     "core-js": "^3.12.1",
@@ -63,9 +60,10 @@
     "moment": "^2.29.4",
     "moment-range": "^4.0.2",
     "ngx-clipboard": "^15.0.0",
+    "ngx-cookie-service": "^15.0.0",
     "ngx-cookieconsent": "^4.0.1",
-    "ngx-csv-parser": "^1.2.0",
-    "ngx-toastr": "^14.0.0",
+    "ngx-csv-parser": "^1.2.2",
+    "ngx-toastr": "^16.1.1",
     "nouislider": "^15.6.1",
     "npm": "^8.11.0",
     "ol": "~6.5.0",
@@ -81,7 +79,7 @@
     "timers": "^0.1.1",
     "tippy.js": "^5.2.1",
     "tslib": "^2.2.0",
-    "typescript": "^4.6.4",
+    "typescript": "^4.9.5",
     "ua-parser-js": "^0.7.28",
     "uuid": "^7.0.3",
     "webpack": "^5.72.1",
@@ -90,20 +88,20 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-eslint/builder": "14.1.2",
-    "@angular-eslint/eslint-plugin": "14.1.2",
-    "@angular-eslint/eslint-plugin-template": "14.1.2",
-    "@angular-eslint/schematics": "14.1.2",
-    "@angular-eslint/template-parser": "14.1.2",
-    "@angular/language-service": "^14.2.8",
+    "@angular-eslint/builder": "15.2.1",
+    "@angular-eslint/eslint-plugin": "15.2.1",
+    "@angular-eslint/eslint-plugin-template": "15.2.1",
+    "@angular-eslint/schematics": "15.2.1",
+    "@angular-eslint/template-parser": "15.2.1",
+    "@angular/language-service": "^15.2.8",
     "@bartholomej/ngx-translate-extract": "^8.0.2",
     "@ngrx/schematics": "^14.3.2",
     "@types/jasmine": "^3.6.11",
     "@types/jasminewd2": "^2.0.8",
     "@types/node": "^12.20.12",
-    "@typescript-eslint/eslint-plugin": "5.37.0",
-    "@typescript-eslint/parser": "5.37.0",
-    "eslint": "^8.26.0",
+    "@typescript-eslint/eslint-plugin": "^5.43.0",
+    "@typescript-eslint/parser": "^5.43.0",
+    "eslint": "^8.28.0",
     "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-jsdoc": "^39.4.0",
     "eslint-plugin-prefer-arrow": "^1.2.3",
