@@ -4,7 +4,7 @@ import { CommonModule } from '@angular/common';
 import { MatSharedModule } from '@shared';
 
 import { BaselineChartComponent } from './baseline-chart.component';
-import { AngularResizeEventModule } from 'angular-resize-event';
+import { ResizedEventModule } from '@directives/resized.directive';
 import { TranslateModule } from "@ngx-translate/core";
 
 
@@ -13,7 +13,7 @@ import { TranslateModule } from "@ngx-translate/core";
   imports: [
     CommonModule,
     MatSharedModule,
-    AngularResizeEventModule,
+    ResizedEventModule,
     TranslateModule,
   ],
   exports: [
