@@ -1,4 +1,4 @@
-import { Component, OnDestroy, OnInit } from '@angular/core';
+import {Component, EventEmitter, OnDestroy, OnInit, Output} from '@angular/core';
 import { Store } from '@ngrx/store';
 import { AppState } from '@store';
 import * as userStore from '@store/user';
@@ -11,6 +11,8 @@ import {
 import { Hyp3Service, ThemingService } from '@services';
 import { SubSink } from 'subsink';
 import { take } from 'rxjs';
+import { TranslateService } from "@ngx-translate/core";
+import { AsfLanguageService } from "@services/asf-language.service";
 
 @Component({
   selector: 'app-preferences',
@@ -18,6 +20,8 @@ import { take } from 'rxjs';
   styleUrls: ['./preferences.component.scss']
 })
 export class PreferencesComponent implements OnInit, OnDestroy {
+  @Output() selectedChange = new EventEmitter<string>();
+
   public datasets = datasetList;
   public defaultMaxResults: number;
   public defaultMapLayer: MapLayerTypes;
@@ -26,10 +30,6 @@ export class PreferencesComponent implements OnInit, OnDestroy {
   public defaultProductTypes: ProductType[];
   public hyp3BackendUrl: string;
   public defaultLanguage: string;
-
-  public defaultGeoSearchFiltersID;
-  public defaultBaselineSearchFiltersID;
-  public defaultSBASSearchFiltersID;
 
   public maxResults = [250, 1000, 5000];
   public mapLayerTypes = MapLayerTypes;
@@ -59,6 +59,9 @@ export class PreferencesComponent implements OnInit, OnDestroy {
     private store$: Store<AppState>,
     private hyp3: Hyp3Service,
     private themeService: ThemingService,
+    public translate: TranslateService,
+    public language: AsfLanguageService,
+
   ) { }
 
   ngOnInit() {
@@ -141,7 +144,8 @@ export class PreferencesComponent implements OnInit, OnDestroy {
   }
 
   public onChangeDefaultLanguage(language: string): void {
-    this.defaultLanguage = language;
+    this.language.setCurrent(language);
+    this.defaultLanguage = language
     this.saveProfile();
   }
 
@@ -197,6 +201,10 @@ export class PreferencesComponent implements OnInit, OnDestroy {
     });
 
     this.store$.dispatch(action);
+  }
+
+  onCloseDownloadQueue() {
+    this.dialogRef.close();
   }
 
   ngOnDestroy() {
