@@ -27,7 +27,7 @@ import { ResizableModule } from 'angular-resizable-element';
 import { DragDropModule } from '@angular/cdk/drag-drop';
 import { ProjectNameSelectorModule } from '@components/shared/selectors/project-name-selector';
 import { ConfirmationComponent } from './confirmation/confirmation.component';
-import { AngularResizeEventModule } from 'angular-resize-event';
+import { ResizedEventModule } from '@directives/resized.directive';
 import { CreateSubscriptionModule } from '../../header/create-subscription';
 import {DocsModalModule} from '@components/shared/docs-modal';
 import {Hyp3UrlModule} from '@components/shared/hyp3-url/hyp3-url.module';
@@ -59,7 +59,7 @@ import { SharedModule } from '@shared';
         ProcessingOptionsModule,
         MatDialogModule,
         ResizableModule,
-        AngularResizeEventModule,
+        ResizedEventModule,
         DragDropModule,
         MatTabsModule,
         MatMenuModule,
