@@ -51,7 +51,7 @@ export class MobileResultsMenuComponent implements OnInit, OnDestroy {
       this.store$.select(searchStore.getSearchType).subscribe(
         searchType => {
           this.searchType = searchType;
-          this.view = searchType === 'SBAS_SEARCH' ?
+          this.view = searchType === SearchType.SBAS ?
                       MobileViews.SBAS : MobileViews.LIST;
         }
       )
@@ -108,7 +108,6 @@ export class MobileResultsMenuComponent implements OnInit, OnDestroy {
     this.view = MobileViews.SBAS;
   }
   public isGraphDisconnected(disconnect: boolean) {
-    console.log(disconnect);
     this.isDisconnected = disconnect;
   }
   ngOnDestroy() {
