@@ -1,8 +1,10 @@
 import '@formatjs/intl-displaynames/polyfill'
-import '@formatjs/intl-displaynames/locale-data/en' // locale-data for en
+import '@formatjs/intl-displaynames/locale-data/en'
+import '@formatjs/intl-displaynames/locale-data/es'
 
-import {Component, EventEmitter, Input, OnInit, Output} from '@angular/core';
-import {TranslateService} from "@ngx-translate/core";
+import { Component, Input, OnInit } from '@angular/core';
+import { TranslateService } from "@ngx-translate/core";
+import { AsfLanguageService } from "@services/asf-language.service";
 
 @Component({
   selector: 'app-language-selector',
@@ -12,11 +14,12 @@ import {TranslateService} from "@ngx-translate/core";
 export class LanguageSelectorComponent implements OnInit {
   @Input() header: boolean;
   @Input() selected: string;
-  @Output() selectedChange = new EventEmitter<string>();
-  public languageNamesInEnglish = new Intl.DisplayNames(['en'], { type: 'language' });
+
+  // public languageNamesInEnglish = new Intl.DisplayNames(['en'], { type: 'language' });
 
   constructor(
     public translate: TranslateService,
+    public language: AsfLanguageService,
   ) { }
 
   ngOnInit(): void {
@@ -24,17 +27,5 @@ export class LanguageSelectorComponent implements OnInit {
       this.header = false;
     }
   }
-  public languageName( langName : string ) {
-    return this.languageNamesInEnglish.of( langName );
-  }
 
-  public onSelectionChange(language: string): void {
-    window.dataLayer = window.dataLayer || [];
-    window.dataLayer.push({
-      'event': 'language-selected',
-      'language': language,
-    });
-    this.translate.use(language)
-    this.selectedChange.emit(language);
-  }
 }
