@@ -1,9 +1,11 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { LanguageSelectorComponent } from "./language-selector.component";
-import { SharedModule } from "@shared";
+import {MatSharedModule, SharedModule} from "@shared";
 import {MatFormFieldModule} from "@angular/material/form-field";
 import {MatSelectModule} from "@angular/material/select";
+import {MatIconModule} from "@angular/material/icon";
+import {MatMenuModule} from "@angular/material/menu";
 
 
 @NgModule({
@@ -14,7 +16,10 @@ import {MatSelectModule} from "@angular/material/select";
     CommonModule,
     SharedModule,
     MatFormFieldModule,
-    MatSelectModule
+    MatSelectModule,
+    MatIconModule,
+    MatSharedModule,
+    MatMenuModule
   ],
   exports: [
     LanguageSelectorComponent
