@@ -4,7 +4,7 @@ export const beta = {
   id: 'SENTINEL-1 INTERFEROGRAM (BETA)',
   name: 'ARIA S1 GUNW',
   subName: 'NISAR-format',
-  beta: true,
+  beta: false,
   properties: [
     Props.DATE,
     Props.BEAM_MODE,
@@ -17,7 +17,7 @@ export const beta = {
   apiValue: { asfplatform: 'SENTINEL-1 INTERFEROGRAM (BETA)' },
   date: { start: new Date(2014, 3, 25) },
   infoUrl: 'https://asf.alaska.edu/data-sets/derived-data-sets/sentinel-1-interferograms/',
-  citationUrl: 'https://asf.alaska.edu/how-to/data-basics/datasets-available-from-asf-sar-daac/',
+  citationUrl: 'https://asf.alaska.edu/data-sets/derived-data-sets/sentinel-1-interferograms/#citation',
   frequency: 'C-Band',
   source: {
     name: 'ARIA-JPL',
