@@ -1,48 +1,57 @@
 import { Pipe, PipeTransform } from '@angular/core';
-
+import { TranslateService } from "@ngx-translate/core";
 import * as moment from 'moment';
 
-
 @Pipe({
-  name: 'fullDate'
+  name: 'fullDate',
+  pure: false
 })
 export class FullDatePipe implements PipeTransform {
+  constructor(private translateService: TranslateService) {}
   transform(date: Date | moment.Moment): string {
     const dateUtc = moment.utc(date);
-
-    return dateUtc.format('MMMM DD YYYY HH:mm:ss') + 'Z';
+    dateUtc.locale(this.translateService.currentLang);
+    return dateUtc.format('LL, HH:mm:ss') + 'Z';
   }
 }
 
 
 @Pipe({
-  name: 'shortDate'
+  name: 'shortDate',
+  pure: false
 })
 export class ShortDatePipe implements PipeTransform {
+  constructor(private translateService: TranslateService) {
+  }
 
   transform(date: Date | moment.Moment): string {
     const dateUtc = moment.utc(date);
-
+    dateUtc.locale(this.translateService.currentLang);
     return dateUtc.format('MMM DD YYYY');
   }
 }
 
 @Pipe({
-  name: 'shortDateTime'
+  name: 'shortDateTime',
+  pure: false
 })
 export class ShortDateTimePipe implements PipeTransform {
-
+  constructor(private translateService: TranslateService) {}
   transform(date: Date | moment.Moment): string {
     const dateUtc = moment.utc(date);
-
-    return dateUtc.format('MM/DD/YY, HH:mm:ss') + 'Z';
+    dateUtc.locale(this.translateService.currentLang);
+    return dateUtc.format('L, HH:mm:ss') + 'Z';
   }
 }
 
 @Pipe({
-  name: 'shortDateSeason'
+  name: 'shortDateSeason',
+  pure: false
 })
 export class ShortDateSeasonPipe implements PipeTransform {
+
+  constructor(private translateService: TranslateService) {}
+
   transform(dayOfYear: number): string {
     const date = new Date();
     date.setFullYear(2019);
@@ -54,9 +63,8 @@ export class ShortDateSeasonPipe implements PipeTransform {
     const dayNumMilli = dayOfYear * dayMilli;
     date.setTime(timeOfFirst + dayNumMilli);
 
-    return  date.toLocaleDateString('en-US', {
+    return  date.toLocaleDateString(this.translateService.currentLang, {
       month: <'numeric'>'short', day: 'numeric'
     });
   }
 }
-
