@@ -1,15 +1,27 @@
 import '@formatjs/intl-displaynames/polyfill'
-import '@formatjs/intl-displaynames/locale-data/en' // locale-data for en
-import '@formatjs/intl-displaynames/locale-data/fr' // locale-data for fr
+import '@formatjs/intl-displaynames/locale-data/en'
+import '@formatjs/intl-displaynames/locale-data/es'
 
-import {NgModule} from '@angular/core';
-import {TranslateModule} from '@ngx-translate/core';
-import {CommonModule} from "@angular/common";
+import { NgModule } from '@angular/core';
+// import { TranslateModule}  from '@ngx-translate/core';
+import { CommonModule } from "@angular/common";
 
-@NgModule({
+import { TranslateLoader, TranslateModule } from '@ngx-translate/core';
+import { TranslateHttpLoader } from "@ngx-translate/http-loader";
+import { HttpClient } from "@angular/common/http";
+
+export function HttpLoaderFactory(http: HttpClient) {
+  return new TranslateHttpLoader(http);
+}@NgModule({
   imports: [
     CommonModule,
-    TranslateModule
+    TranslateModule.forChild({
+      loader: {
+        provide: TranslateLoader,
+        useFactory: (HttpLoaderFactory),
+        deps: [ HttpClient ]
+      }
+    })
   ],
   exports: [
     TranslateModule,
